requirements.txt
# Core scientific stack
numpy
pandas
matplotlib

# Hugging Face dataset loader (used for autogluon/fev_datasets)
datasets

# Amazon Chronos / Chronos-2 library (correct package)
chronos-forecasting

# Chronos-forecasting core dependencies (explicit for reproducibility)
torch
transformers
accelerate
einops
scikit-learn

